requirements.txt
fastapi>=0.115,<1.0
pdf2docx>=0.5.8,<1.0
python-dotenv>=1.0,<2.0
python-multipart>=0.0.9,<1.0
requests>=2.32,<3.0
uvicorn[standard]>=0.30,<1.0
